chore: move out viewer
